requirements.txt
openai
python-dotenv
PyPDF2  # for handling PDF files
streamlit  # for building the front-end interface
